+++ FuelFans/Models/Reward.cs
@@ -2,7 +2,7 @@
 {
     public class Reward
     {
-        public int Number { get; set; }
+        public decimal Number { get; set; }
         public string Name { get; set; }
         public string ImageUrl { get; set; }
     }
